Fix aioble on macOS Catalina (#27)
In macOS Catalina the behavior of CBCentralManager has changed so that
it calls it's delegate before returning from it's initializer. This
causes a problem since the delegate implementation in centralmanager.py
accesses self._cbmanager which hasn't been set because the initializer
hasn't returned yet.

This patch makes the centralmanager use either the passed in cbmanager object from the delegate call and then fall back to using the self._cbmanager member variable if the parameter is None.